Remove outdated template backups
The way the plugin works right now, once a template has been used for a
backup once, it is never updated until the caches are manually flushed.
This is not ideal for long-running servers whose caches might be very
long lived or when developing templates. This change adds listeners to
all course-related events and deletes the old backup file when changes
are made.